+++ packages/lesswrong/components/review/ReviewVotingButtons.tsx
@@ -6,9 +6,8 @@ import forumThemeExport from '../../themes/forumTheme';
 import { DEFAULT_QUALITATIVE_VOTE } from '../../lib/collections/reviewVotes/schema';
 import { AnalyticsContext } from '../../lib/analyticsEvents';
 import { useCurrentUser } from '../common/withUser';
-import { eligibleToNominate } from '../../lib/reviewUtils';
+import { eligibleToNominate, getCostData } from '../../lib/reviewUtils';
 import { SyntheticQualitativeVote } from './ReviewVotingPage';
-import round from 'lodash/round';
 
 const downvoteColor = "rgba(125,70,70, .87)"
 const upvoteColor = forumTypeSetting.get() === "EAForum" ? forumThemeExport.palette.primary.main : "rgba(70,125,70, .87)"
@@ -53,61 +52,6 @@ const styles = (theme: ThemeType) => ({
   7: { color: upvoteColor},
 })
 
-const getPointsFromCost = (cost) => {
-  // the formula to quadratic cost from a number of points is (n^2 + n)/2
-  // this uses the inverse of that formula to take in a cost and output a number of points
-  return (-1 + Math.sqrt((8 * cost)+1)) / 2
-}
-
-const getLabelFromCost = (cost) => {
-  // rounds the points to 1 decimal for easier reading
-  return round(getPointsFromCost(cost), 1)
-}
-
-export const getCostData = ({costTotal=500}:{costTotal?:number}) => {
-  // This 
-  const divider = costTotal > 500 ? costTotal/500 : 1
-  return ({
-    0: {label: null, cost: 0, tooltip: null},
-    1: { label: `-${getLabelFromCost(45/divider)}`, cost: 45, tooltip: 
-      <div>
-        <p>Highly misleading, harmful, or unimportant.</p>
-        <p><em>Costs 45 points</em></p>
-      </div>},
-    2: { label: `-${getLabelFromCost(10/divider)}`, cost: 10, tooltip: 
-    <div>
-      <p>Very misleading, harmful, or unimportant.</p>
-      <p><em>Costs 10 points</em></p>
-    </div>},
-    3: { label: `-${getLabelFromCost(1/divider)}`, cost: 1, tooltip: 
-    <div>
-      <p>Misleading, harmful or unimportant.</p>
-      <p><em>Costs 1 point</em></p>
-    </div>},
-    4: { label: `0`, cost: 0, tooltip: 
-    <div>
-      <p>No strong opinion on this post,</p>
-      <p><em>Costs 0 points</em></p>
-    </div>},
-    5: { label: `${getLabelFromCost(1/divider)}`, cost: 1, tooltip: 
-    <div>
-      <p>Good</p>
-      <p><em>Costs 1 point</em></p>
-    </div>},
-    6: { label: `${getLabelFromCost(10/divider)}`, cost: 10, tooltip: 
-    <div>
-      <p>Quite important</p>
-      <p><em>Costs 10 points</em></p>
-    </div>},
-    7: { label: `${getLabelFromCost(45/divider)}`, cost: 45, tooltip: 
-    <div>
-      <p>Extremely important</p>
-      <p><em>Costs 45 points</em></p>
-    </div>},
-  })
-}
-
-
 const ReviewVotingButtons = ({classes, post, dispatch, currentUserVote, costTotal}: {classes: ClassesType, post: PostsMinimumInfo, dispatch: any, currentUserVote: SyntheticQualitativeVote|null, costTotal?: number}) => {
   const { LWTooltip } = Components
 
@@ -135,8 +79,8 @@ const ReviewVotingButtons = ({classes, post, dispatch, currentUserVote, costTota
   return <AnalyticsContext pageElementContext="reviewVotingButtons">
     <div className={classes.root}>
         {[1,2,3,4,5,6,7].map((i) => {
-          return <LWTooltip title={getCostData({})[i].tooltip} 
-          key={`${getCostData({})[i]}-${i}`}>
+          return <LWTooltip title={getCostData({costTotal})[i].tooltip} 
+          key={`${getCostData({costTotal})[i]}-${i}`}>
             <span
                 className={classNames(classes.button, classes[i], {
                   [classes.selectionHighlight]:selection === i && !isDefaultVote,
@@ -144,7 +88,7 @@ const ReviewVotingButtons = ({classes, post, dispatch, currentUserVote, costTota
                 })}
                 onClick={createClickHandler(i)}
               >
-              {getCostData({costTotal:costTotal})[i].label}
+              {getCostData({costTotal})[i].label}
             </span>
           </LWTooltip>
         })}
