+++ packages/lesswrong/lib/reviewUtils.tsx
@@ -1,3 +1,5 @@
+import React from 'react';
+import round from "lodash/round"
 import moment from "moment"
 import { forumTypeSetting } from "./instanceSettings"
 import { annualReviewEnd, annualReviewNominationPhaseEnd, annualReviewReviewPhaseEnd, annualReviewStart } from "./publicSettings"
@@ -66,4 +68,58 @@ export const currentUserCanVote = (currentUser: UsersCurrent|null) => {
   if (isLWForum && new Date(currentUser.createdAt) > new Date(`${REVIEW_YEAR+1}-01-01`)) return false
   if (isEAForum && new Date(currentUser.createdAt) > new Date(annualReviewStart.get())) return false
   return true
+}
+
+const getPointsFromCost = (cost) => {
+  // the formula to quadratic cost from a number of points is (n^2 + n)/2
+  // this uses the inverse of that formula to take in a cost and output a number of points
+  return (-1 + Math.sqrt((8 * cost)+1)) / 2
+}
+
+const getLabelFromCost = (cost) => {
+  // rounds the points to 1 decimal for easier reading
+  return round(getPointsFromCost(cost), 1)
+}
+
+export const getCostData = ({costTotal=500}:{costTotal?:number}) => {
+  const divider = costTotal > 500 ? costTotal/500 : 1
+  const dividerWarning = divider && <div>Your vote is downweighted because you spent 500+ points</div>
+  return ({
+    0: {label: null, cost: 0, tooltip: null},
+    1: { label: `-${getLabelFromCost(45/divider)}`, cost: 45, tooltip: 
+      <div>
+        <p>Highly misleading, harmful, or unimportant.</p>
+        <p><em>Costs 45 points{dividerWarning}</em></p>
+      </div>},
+    2: { label: `-${getLabelFromCost(10/divider)}`, cost: 10, tooltip: 
+    <div>
+      <p>Very misleading, harmful, or unimportant.</p>
+      <p><em>Costs 10 points{dividerWarning}</em></p>
+    </div>},
+    3: { label: `-${getLabelFromCost(1/divider)}`, cost: 1, tooltip: 
+    <div>
+      <p>Misleading, harmful or unimportant.</p>
+      <p><em>Costs 1 point{dividerWarning}</em></p>
+    </div>},
+    4: { label: `0`, cost: 0, tooltip: 
+    <div>
+      <p>No strong opinion on this post,</p>
+      <p><em>Costs 0 points{dividerWarning}</em></p>
+    </div>},
+    5: { label: `${getLabelFromCost(1/divider)}`, cost: 1, tooltip: 
+    <div>
+      <p>Good</p>
+      <p><em>Costs 1 point{dividerWarning}</em></p>
+    </div>},
+    6: { label: `${getLabelFromCost(10/divider)}`, cost: 10, tooltip: 
+    <div>
+      <p>Quite important</p>
+      <p><em>Costs 10 points{dividerWarning}</em></p>
+    </div>},
+    7: { label: `${getLabelFromCost(45/divider)}`, cost: 45, tooltip: 
+    <div>
+      <p>Extremely important</p>
+      <p><em>Costs 45 points{dividerWarning}</em></p>
+    </div>},
+  })
 }
